Move garbage collection roots into new "libroots"
The new "libroots" library intentionally does not rely on other parts of
the nix codebase, so as to allow creating a minimal daemon to find roots
while the rest of nix is left unprivileged.
