+++ packaging/components.nix
@@ -411,6 +411,8 @@ in
   nix-flake-c = callPackage ../src/libflake-c/package.nix { };
   nix-flake-tests = callPackage ../src/libflake-tests/package.nix { };
 
+  nix-roots = callPackage ../src/libroots/package.nix { };
+
   nix-main = callPackage ../src/libmain/package.nix { };
   nix-main-c = callPackage ../src/libmain-c/package.nix { };
