+++ src/libroots/.version
@@ -1,0 +1,1 @@
+../../.version
