+++ src/libroots/include/nix/roots/meson.build
@@ -1,0 +1,5 @@
+# Public headers directory
+
+include_dirs = [ include_directories('../..') ]
+
+headers = files('roots.hh')
