+++ src/libroots/nix-meson-build-support
@@ -1,0 +1,1 @@
+../../nix-meson-build-support
